Force alignment in ARM architectures
The "opaque but sized" structures (dill_*_storage) may cause data (in particular
pointers) not to be aligned. This is not a problem in x86_64, but in ARM it causes
loads and stores to be unpredictable (i.e. store wrong data). Then the code will
try to invoke a function from a wrong pointer and jumps to nowhere, causing a segfault.