epoll: avoid fd leaks to child processes by using epoll_create1.
There is no use case in which one would need to share libdill's epoll
file descriptor with a child process. Even if it was innocuous, it makes
automated checking difficult as one has to consider this special case.

Since libdill does not provide access to the internal epoll fd, the only
way to fix this is to change the source code itself.