dill_fd_own: preserve fd-flags (CLOEXEC) after dup.
The IPC and TPC subsystems call dill_fd_own which does not preserve the
file descriptor flags. This means that even if the user creates the sockets
with the proper flags, when they use "fromfd" the file descriptors will still be
leaked to child processes.

The implementation tries to use F_DUPFD_CLOEXEC if available.